Add files via upload
Distribution data visualizations from the project. What comes to the attention is how marginalized municipalities in Mexico are not only, as is evident, more vulnerable to climate related events, but also have and increased risk rate of exposure to at leats one of these events: landslides. This is probably due to the fact that marginalized communities in Mexico live on more uneven land and with substantially different altitudes within its territory.